feat(vm): better modularity for this crate
